+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 * Remove the portfinder dependency and let ADB find an unused port instead
+* Improve the WebView detection in the ADB client bridge
 
 ## 1.0.1 - 2018-09-01
 * Rename the `adbExecutable` configuration parameter